# Remove IAvaloniaPropertyVisitor due to generic virtual methods.