… src/Avalonia.Base/Utilities/IAvaloniaPropertyVisitor.cs