## What does the pull request do?

Removes `IAvaloniaPropertyVisitor` which was based around using generic virtual methods, which we're trying to remove. It was only used internally in creating `ISetterInstance`s so adds a virtual method to `AvaloniaProperty` to do this explicitly without generic virtual methods.

This PR includes the changes from #8021 so until that is merged you can see the changes specific to this PR in the commit [Remove IAvaloniaPropertyVisitor.](https://github.com/AvaloniaUI/Avalonia/commit/6df672e4c078bae8dd0259825032af68824f3f08)