@avaloniaui-team 2022-04-20
You can test this PR using the following package version. `0.10.999-cibuild0020067-beta`. (feed url: https://nuget.avaloniaui.net/repository/avalonia-all/index.json) [PRBUILDID]