@grokys 2022-04-21
Benchmarks as compared with #7980 and filtered for noise:

```
| Faster                                                                           | base/diff | Base Median (ns) | Diff Median (ns) | Modality  |
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| ---------- |
| Avalonia.Benchmarks.Layout.ControlsBenchmark.CreateTextBox                       |      2.10 |       2782200.00 |       1325559.18 | multimodal|
```

Only one benchmark has changed, but it's 2x faster.